dateOptionsとtagOptionsをutilsに分離
