+++ src/components/table/tagCell.tsx
@@ -4,37 +4,13 @@ import { jsx } from '@emotion/core';
 import { Table, Label, Dropdown, DropdownProps } from 'semantic-ui-react';
 import dayjs from 'dayjs';
 import { StoreExpense } from '../../interfaces';
+import { tagOptions } from '../../utils/utils';
 
 interface TagCellProps {
   expense?: StoreExpense;
   isEditable?: boolean;
   handleChangeTag?: (e: FormEvent, { key }: DropdownProps) => void;
 }
-
-export const tagOptions = [
-  { key: 0, text: 'その他', value: 'その他', icon: 'tag', color: '#838383', colorLabel: 'grey' },
-  { key: 1, text: '食費', value: '食費', icon: 'food', color: '#049C94', colorLabel: 'teal' },
-  { key: 2, text: '家賃', value: '家賃', icon: 'home', color: '#5929BB', colorLabel: 'violet' },
-  {
-    key: 3,
-    text: '電気代',
-    value: '電気代',
-    icon: 'power cord',
-    color: '#EBAE00',
-    colorLabel: 'yellow',
-  },
-  { key: 4, text: '水道代', value: '水道代', icon: 'bath', color: '#1778C1', colorLabel: 'blue' },
-  { key: 5, text: 'ガス代', value: 'ガス代', icon: 'fire', color: '#D11A1A', colorLabel: 'red' },
-  { key: 6, text: '電話', value: '電話', icon: 'phone', color: '#F36203', colorLabel: 'orange' },
-  {
-    key: 7,
-    text: '交通費',
-    value: '交通費',
-    icon: 'subway',
-    color: '#13AB38',
-    colorLabel: 'green',
-  },
-];
 
 export const TagCellComponent: FC<TagCellProps> = ({
   expense = {
